Move Buttons component to /layouts
